hal/string: change declaration of hal_i2s
also:
 - fix hal_strncpy on ia32 and riscv

JIRA: PD-210

+++ hal/string.h
@@ -44,7 +44,7 @@ extern char *hal_strcpy(char *dest, const char *src);
 extern char *hal_strncpy(char *dest, const char *src, size_t n);
 
 
-extern int hal_i2s(char *prefix, char *s, unsigned int i, unsigned char b, char zero);
+extern unsigned long hal_i2s(char *prefix, char *s, unsigned long i, unsigned char b, char zero);
 
 
 #endif
